refactored search calls page following nextjs struct
