+++ src/App.tsx
@@ -98,15 +98,6 @@ function App({
 
   const facilities = useSelector(selectAllFacilities);
 
-  // useEffect(() => {
-  //   localStorage.removeItem("debug");
-  //   (async () => {
-  //     try {
-  //       await fetchFacilities();
-  //     } catch (err) {}
-  //   })();
-  // }, [fetchFacilities]);
-
   useEffect(() => {
     if (isAuthenticated) fetchFacilities();
   }, [isAuthenticated, fetchFacilities]);
