+++ src/components/SearchCalls/index.tsx
@@ -1,0 +1,3 @@
+import SearchCalls from "./SearchCalls";
+
+export default SearchCalls;
